Struct wasmer_engine_native::engine::NativeEngineInner

pub struct NativeEngineInner {
    compiler: Option<Box<dyn Compiler>>,
    features: Features,
    signatures: SignatureRegistry,
    prefixer: Option<Box<dyn Fn(&[u8]) -> String + Send>>,
    is_cross_compiling: bool,
    linker: Linker,
    libraries: Vec<Library>,
}

The inner contents of NativeEngine

Fields
compiler: Option<Box<dyn Compiler>>

The compiler

features: Features

The WebAssembly features to use

signatures: SignatureRegistry

The signature registry is used mainly to operate with trampolines performantly.

prefixer: Option<Box<dyn Fn(&[u8]) -> String + Send>>

The prefixer returns the a String to prefix each of the functions in the shared object generated by the NativeEngine, so we can assure no collisions.

is_cross_compiling: bool

Whether the native engine will cross-compile.

linker: Linker

The linker to use.

libraries: Vec<Library>

List of libraries loaded by this engine.

Implementations

impl NativeEngineInner

pub fn compiler(&self) -> Result<&dyn Compiler, CompileError>

Gets the compiler associated to this engine.

pub(crate) fn get_prefix(&self, bytes: &[u8]) -> String

pub(crate) fn features(&self) -> &Features

pub fn validate<'data>(&self, data: &'data [u8]) -> Result<(), CompileError>

Validate the module

pub fn signatures(&self) -> &SignatureRegistry

Shared signature registry.

pub(crate) fn is_cross_compiling(&self) -> bool

pub(crate) fn linker(&self) -> Linker

pub(crate) fn add_library(&mut self, library: Library)